+++ client/lib/js/adapters.ts
@@ -176,6 +176,14 @@ namespace ABeamer {
     getProp(name: PropName, args?: ABeamerArgs): PropValue;
     setProp(name: PropName, value: PropValue, args?: ABeamerArgs): void;
     waitFor?(waitFor: WaitFor, onDone: DoneFunc, args?: ABeamerArgs): void;
+
+
+    /**
+     * Called after the frame is rendered, and before moves to the next frame.
+     * This method is called even if no property changed.
+     * It's an optional method, but future version might require its implementation.
+     */
+    frameRendered?(args?: ABeamerArgs): void;
   }
 
 
@@ -207,13 +215,32 @@ namespace ABeamer {
 
     props: AnyParams = {};
     selector: string;
+    propsChanged: boolean = false;
 
     onAnimateProp: (name: PropName, value: PropValue) => void;
+    onAnimateProps: (args?: ABeamerArgs) => void;
 
 
+    /**
+     * Called after property value changed.
+     * Use this method instead animateProps, if the rendering should be right after
+     * each property is updated, otherwise use animateProps.
+     */
     animateProp(name: PropName, value: PropValue): void {
       if (this.onAnimateProp) {
         this.onAnimateProp(name, value);
+      }
+    }
+
+
+    /**
+     * Called after actions from the frame are rendered, and if at least one property changed.
+     * Use this method instead animateProp, if the animation has multiple virtual properties and
+     * each animation can be done after all are updated.
+     */
+    animateProps(args?: ABeamerArgs): void {
+      if (this.onAnimateProps) {
+        this.onAnimateProps(args);
       }
     }
 
@@ -226,7 +253,16 @@ namespace ABeamer {
     setProp(name: PropName, value: PropValue): void {
       this.props[name] = value;
       if (name !== 'uid') {
+        this.propsChanged = true;
         this.animateProp(name, value);
+      }
+    }
+
+
+    frameRendered(args?: ABeamerArgs) {
+      if (this.propsChanged) {
+        this.animateProps(args);
+        this.propsChanged = false;
       }
     }
   }
@@ -408,6 +444,7 @@ namespace ABeamer {
     abstract getProp(name: PropName, args?: ABeamerArgs): PropValue;
     abstract setProp(name: PropName, value: PropValue, args?: ABeamerArgs): void;
     waitFor?(waitItem: WaitFor, onDone: DoneFunc, args?: ABeamerArgs): void { }
+    frameRendered?(args?: ABeamerArgs): void { }
   }
 
   // ------------------------------------------------------------------------
@@ -651,6 +688,10 @@ namespace ABeamer {
 
     waitFor?(waitItem: WaitFor, onDone: DoneFunc, args?: ABeamerArgs): void {
       this.vElement.waitFor(waitItem, onDone, args);
+    }
+
+    frameRendered?(args: ABeamerArgs) {
+      if (this.vElement.frameRendered) { this.vElement.frameRendered(args); }
     }
   }