+++ CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add support of system env IM_CONVERT_BIN to locate imagemagick convert executable.
 - Add `check` command to the abeamer cli.
 - (DEV) Add windows test batch files.
+- Add adapters.frameRendered.
+- Add SimpleVirtualAnimator.animateProps.
 
 ### Fix
 - Fix http-server for URL with path with %20(spaces).